pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codebase-onboard"
version = "1.0.0"
description = "Analyze any codebase and generate comprehensive onboarding guides for new developers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Sagar Jha", email = "sagarjhaa@gmail.com"}
]
keywords = ["developer-tools", "onboarding", "codebase-analysis", "documentation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pathspec>=0.12.0",
]

[project.optional-dependencies]
ai = ["openai>=1.0.0", "httpx>=0.25.0"]
all = ["openai>=1.0.0", "httpx>=0.25.0", "markdown-it-py>=3.0.0"]

[project.scripts]
onboard = "codebase_onboard.cli:cli"

[project.urls]
Homepage = "https://github.com/sagarjhaa/codebase-onboard"
Repository = "https://github.com/sagarjhaa/codebase-onboard"
Issues = "https://github.com/sagarjhaa/codebase-onboard/issues"
